backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
alembic
psycopg2-binary
asyncpg
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv
httpx
pytest
pytest-asyncio
loguru
